Revert "fix(useCookieState): default value is not stored (#2098)"
This reverts commit 1a176f8a7c4e7f3e201562bccd29bb025eeda69b.
