+++ packages/hooks/src/useCookieState/index.ts
@@ -1,8 +1,7 @@
 import Cookies from 'js-cookie';
 import { useState } from 'react';
 import useMemoizedFn from '../useMemoizedFn';
-import useUpdateEffect from '../useUpdateEffect';
-import { isFunction, isString, isUndef } from '../utils';
+import { isFunction, isString } from '../utils';
 
 export type State = string | undefined;
 
@@ -11,48 +10,33 @@ export interface Options extends Cookies.CookieAttributes {
 }
 
 function useCookieState(cookieKey: string, options: Options = {}) {
-  function getDefaultValue() {
-    return isFunction(options?.defaultValue) ? options?.defaultValue() : options?.defaultValue;
-  }
-
-  function setStoredValue(newValue: State, newOptions: Cookies.CookieAttributes = {}) {
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    const { defaultValue, ...restOptions } = newOptions;
-
-    if (isUndef(newValue)) {
-      Cookies.remove(cookieKey);
-    } else {
-      Cookies.set(cookieKey, newValue, restOptions);
-    }
-  }
-
-  function getStoredValue() {
+  const [state, setState] = useState<State>(() => {
     const cookieValue = Cookies.get(cookieKey);
 
     if (isString(cookieValue)) return cookieValue;
 
-    const defaultValue = getDefaultValue();
+    if (isFunction(options.defaultValue)) {
+      return options.defaultValue();
+    }
 
-    setStoredValue(defaultValue);
-
-    return defaultValue;
-  }
-
-  const [state, setState] = useState<State>(() => getStoredValue());
-
-  useUpdateEffect(() => {
-    setState(getStoredValue());
-  }, [cookieKey]);
+    return options.defaultValue;
+  });
 
   const updateState = useMemoizedFn(
     (
       newValue: State | ((prevState: State) => State),
       newOptions: Cookies.CookieAttributes = {},
     ) => {
-      const currentValue = isFunction(newValue) ? newValue(state) : newValue;
+      const { defaultValue, ...restOptions } = { ...options, ...newOptions };
+      const value = isFunction(newValue) ? newValue(state) : newValue;
 
-      setState(currentValue);
-      setStoredValue(currentValue, newOptions);
+      setState(value);
+
+      if (value === undefined) {
+        Cookies.remove(cookieKey);
+      } else {
+        Cookies.set(cookieKey, value, restOptions);
+      }
     },
   );
